requirements.txt
openai
python-dotenv
duckduckgo-search
Pillow
rich
